iris: Fix for PIPE_CAP_SIGNED_VERTEX_BUFFER_OFFSET
This fixes ext_transform_feedback-builtin-varyings gl_Position after the
combination of my transform feedback reworks and my vertex buffer
reworks (?)

+++ src/gallium/drivers/iris/iris_state.c
@@ -2379,7 +2379,7 @@ iris_set_vertex_buffers(struct pipe_context *ctx,
          if (res) {
             vb.BufferSize = res->bo->size;
             vb.BufferStartingAddress =
-               ro_bo(NULL, res->bo->gtt_offset + buffer->buffer_offset);
+               ro_bo(NULL, res->bo->gtt_offset + (int) buffer->buffer_offset);
          } else {
             vb.NullVertexBuffer = true;
          }
